AnalogSource: return a float from read() method
this gives the full resolution when using the ADC backend. It also
adds scaling for the Arduino backend

+++ libraries/AP_AnalogSource/AP_AnalogSource_ADC.cpp
@@ -1,9 +1,9 @@
 
 #include "AP_AnalogSource_ADC.h"
 
-int AP_AnalogSource_ADC::read(void)
+float AP_AnalogSource_ADC::read(void)
 {
-    int fullscale = _adc->Ch(_ch);
-    int scaled = _prescale * fullscale;
+    float fullscale = _adc->Ch(_ch);
+    float scaled = _prescale * fullscale;
     return scaled;
 }
